% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GR_GENE_MODEL}
\alias{GR_GENE_MODEL}
\title{germline small variant object}
\format{
## `GR_GENE_MODEL`
GRanges object
}
\usage{
data(GR_GENE_MODEL)
}
\value{
Object containing gene model of hg38
}
\description{
germline small variant object
}
\keyword{datasets}
