% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GR_GERM_SMALL_VARS}
\alias{GR_GERM_SMALL_VARS}
\title{germline small variant object}
\format{
## `GR_SOM_SMALL_VARS`
GRanges object
}
\usage{
data(GR_SOM_SMALL_VARS)
}
\description{
germline small variant object
}
\keyword{datasets}
