% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GR_SCNA}
\alias{GR_SCNA}
\title{copynumber object}
\format{
## `GR_SCNA`
GRanges object
}
\usage{
data(GR_SCNA)
}
\value{
Object containing somatic copy number abberations (sCNAs) of SeqC2 example case
}
\description{
copynumber object
}
\keyword{datasets}
