% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{ZP_ov}
\alias{ZP_ov}
\title{accesor for ZygoistyPredictor runs. Prints an overview about the run}
\usage{
ZP_ov(fp)
}
\arguments{
\item{fp}{full prediction (output of predict_zygoisty())}
}
\description{
accesor for ZygoistyPredictor runs. Prints an overview about the run
}
