% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{gene_ov}
\alias{gene_ov}
\title{accesor for gene predictions printing detailed info about how a gene status 
was assigned}
\usage{
gene_ov(fp, inp_gene, n = 20)
}
\arguments{
\item{fp}{full prediction (output of predict_zygoisty())}

\item{inp_gene}{name of gene that should be printed with detailed information}

\item{n}{max number of rows to print, as some gene status depend on loads of
phasing results#'}
}
\description{
accesor for gene predictions printing detailed info about how a gene status 
was assigned
}
