% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GR_SOM_SMALL_VARS}
\alias{GR_SOM_SMALL_VARS}
\title{somatic small variant object}
\format{
## `GR_GERM_SMALL_VARS`
GRanges object
}
\usage{
GR_SOM_SMALL_VARS
}
\description{
somatic small variant object
}
\keyword{datasets}
